test(dump): unify dump-test helpers, re-drive workload while polling
